Contact Reception
I added in the layout for the contacts_screen as well as the
functionality to receive all contacts from the phone into a list of
Contact objects, which are simply two Strings. One for a name, and one
for a phone number. The list can then be accessed in the
SpecialTextConfig_Screen.java class as it is now inheriting from the
Contacts_Screen.java class.